feat(qwen36): affine embedding kernel, the three missing int4 variants, f16 MoE output
Three gaps a multi-agent read of the engine surfaced, each of which fails
silently rather than loudly.

1. embedding_affine.wgsl. embedding.wgsl is hard-symmetric — (nibble-7)*scale
   over groups of 32, with no bias binding at all — and the MLX embedding table
   is affine over groups of 64. Binding one to the other is IN BOUNDS and
   ERROR-FREE: the buffers are the right size, the scale index just lands in the
   wrong group and every value is off by its bias. Wrong token vectors, at the
   very top of the forward pass, with nothing downstream pointing back. New
   kernel plus a qwen36embed bundle checking five token ids (including 0 and
   vocab-1, where a row-stride off-by-one shows first) against mlx's own
   dequantize: 7.58e-4.

2. INT4_MATMUL_VARIANTS gains int4_matmul_f32_sg_affine,
   int4_matmul_f32_tiled_affine and int4_matmul_tiled_affine_moe. The last was
   generator-only, constructed ad hoc inside the tests. The f32 siblings are not
   optional extras: variants.ts's resolveMatmul only returns a tiled pipeline
   when BOTH the f16 and f32 forms exist, so without them an untied lm_head
   falls to the scalar kernel at 248320 workgroups — past the 65535
   per-dimension limit — where the tiled one is 62080 and clears it.

3. moe_combine now writes f16, not f32. add_norm reads f16, and WebGPU does not
   type-check storage element types at bind time, so an f32 buffer there is
   accepted and reinterpreted as twice as many f16s. The whole-layer test was
   doing that final residual add ON THE HOST — its own comment said so — which
   means no GPU path from the MoE block into the residual was validated. It now
   runs through add_norm on the GPU, so the seam is covered.

Fixing (3) turned up a second bug in the test itself, worth recording: add_norm's
grid is the TOKEN count (base = batch * D), not ceil(D/256). Dispatching D/256
workgroups for one token runs seven extra workgroups whose reads fall off the end
of the buffer — and their out-of-bounds WRITES do not vanish, they clamp to the
last valid index. One clobbered element out of 2048 moved a max-normalised error
from 8.12e-4 to 9.14e-2.

The MoE block is 7 dispatches, not the 6 three separate places claimed
(router_logits, router_topk, gate, up, silu_mul, down, combine — counted).

All six real-weight checks pass, worst stage 2.48e-3; 32/32 + 19/19 + 26/26
kernels; repack still byte-identical.

+++ src/compiler/shaders/moe_combine.wgsl
@@ -6,14 +6,18 @@
 // sigmoid of its gate there), so there is no special case here — which is the
 // whole point of stacking it into the expert tensors.
 //
-// Accumulates in f32: K+1 f16 terms of similar magnitude summed in f16 loses
-// bits the block cannot spare, and the caller wants f32 for the residual add.
+// Accumulates in f32 and WRITES f16. The accumulator matters — K+1 terms of
+// similar magnitude summed in f16 loses bits the block cannot spare — but the
+// output dtype is fixed by where it goes: add_norm reads f16, exactly as it
+// does for the dense FFN's down projection, so the MoE block is a drop-in for
+// it. Writing f32 here would be silently accepted at bind time (WebGPU does not
+// type-check storage element types) and reinterpreted as twice as many f16s.
 //
 // Grid: ceil(N / 256) workgroups.
 
 enable f16;
 
-@group(0) @binding(0) var<storage, read_write> out_buf : array<f32>;   // [N]
+@group(0) @binding(0) var<storage, read_write> out_buf : array<f16>;   // [N]
 @group(0) @binding(1) var<storage, read> slots : array<f16>;           // [SLOTS, N]
 @group(0) @binding(2) var<storage, read> score : array<f32>;           // [SLOTS]
 
@@ -32,5 +36,5 @@ fn moe_combine(@builtin(global_invocation_id) gid : vec3<u32>) {
   for (var s : u32 = 0u; s < podArgs.SLOTS; s = s + 1u) {
     acc = acc + score[s] * f32(slots[s * podArgs.N + i]);
   }
-  out_buf[i] = acc;
+  out_buf[i] = f16(acc);
 }
